Add files via upload
Uploading pages that represented the state of the user screen when it is first open. 

Need to add the ability to move between pages on the user panel and other functionality associated with the user page. 

The main element currently missing is the album to click onto other user's profiles and album pages. This is because these pages have not been created yet.